refactor timemap construction
